ashan: [Add] SlimWindows.h to minimize the things that are loaded when #include <Windows.h>.
ashan: [Refactor] Refactored whole code to introduce a new windows framework to register, create and show windows on screen.

ashan: [Refactor] As a part of refactor introduced class, singleton, static class instance, and multiple member functions.

ashan: [Modify] Modified the WinMain() function to have minimal things going inside as a part of the new windows framework refactor.

ashan: [Action] A thing that is missing in current window framework is that it is lacking the ability to spawn multiple windows using the framework. Thus, next immediate step is to make the windows framework to be able to handle multiple windows.