ashan: [Add] Added a timer system that calculates variable delta time (dt), frames per second (fps) and average frames per second (avg fps).
ashan: [Add] Added some codes in different classes to test and make FPS get running on the window title.

ashan: [Issue] Found an issue that is caused by GetMessages and it is not the one mentioned in the tutorial video. Should fix it. Added comments about the issue.

ashan: [Action] Need to do code clean up.

ashan: [Action] I felt the exception handling is not required at the moment. Hence, skipped it but should implement it later.

ashan: [Action] Next immediate action is to introduce an array or any best suited container inside Application class of type Window* to hold instances for all the windows to be created.