Issue with using GetMessage() is fixed now by replacing it with PeekMessage()
ashan: [Fix] Fixed an issue with using GetMessage() and replaced it with PeekMessage() as that is the one we should be using for games and engines.

ashan: [Action] Need to do code clean up.

ashan: [Action] I felt the exception handling is not required at the moment. Hence, skipped it but should implement it later.

ashan: [Action] Next immediate action is to introduce an array or any best suited container inside Application class of type Window* to hold instances for all the windows to be created.